requirements.txt
web3
python-dotenv
requests
fastapi
uvicorn
eth-hash
eth-account
python-multipart
httpx
